+++ gdsfactory/geometry/trim.py
@@ -26,12 +26,20 @@ def trim(
     Useful to get a smaller component from a larger one for simulation.
 
     Args:
-        component: Component(/Reference)
+        component: Component(/Reference).
         domain: list of array-like[N][2] representing the boundary of the component to keep.
         precision: float Desired precision for rounding vertex coordinates.
         return_ports: whether to return the included ports or not. Ports are always renamed to avoid inheritance conflicts.
 
     Returns: New component with layers (and possibly ports) of the component restricted to the domain.
+
+    .. plot::
+      :include-source:
+
+      import gdsfactory as gf
+      c = gf.components.straight_pin(length=10, taper=None)
+      trimmed_c = gf.geometry.trim(component=c, domain=[[0, -5], [0, 5], [5, 5], [5, -5]])
+      trimmed_c.plot_matplotlib()
     """
     domain_shape = gdstk.Polygon(domain)
     c = Component()