+++ gdsfactory/simulation/gmeep/write_sparameters_meep.py
@@ -31,7 +31,7 @@ from gdsfactory.simulation.gmeep.get_simulation import (
 from gdsfactory.technology import LayerStack
 from gdsfactory.typings import ComponentSpec, PathType, Port, PortSymmetries
 
-ncores = multiprocessing.cpu_count()
+core_materials = multiprocessing.cpu_count()
 
 
 def remove_simulation_kwargs(d: Dict[str, Any]) -> Dict[str, Any]:
@@ -139,7 +139,8 @@ def write_sparameters_meep(
     decay_by: float = 1e-3,
     is_3d: bool = False,
     z: float = 0,
-    plot_args: Dict = None,
+    plot_args: Optional[Dict] = None,
+    only_return_filepath_sim_settings=False,
     **settings,
 ) -> Dict:
     r"""Returns Sparameters and writes them to npz filepath.
@@ -308,6 +309,11 @@ def write_sparameters_meep(
     sim_settings["component"] = component.to_dict()
     filepath = pathlib.Path(filepath)
     filepath_sim_settings = filepath.with_suffix(".yml")
+
+    # FIXME: Ideally, we should split sim settings generation from doing the
+    #        simulation... this is a hack.
+    if only_return_filepath_sim_settings:
+        return filepath_sim_settings
 
     # filepath_sim_settings.write_text(OmegaConf.to_yaml(sim_settings))
     # logger.info(f"Write simulation settings to {filepath_sim_settings!r}")