+++ gdsfactory/simulation/modes/tests/test_find_modes_dispersion.py
@@ -6,7 +6,7 @@ from gdsfactory.simulation.modes.find_mode_dispersion import find_mode_dispersio
 
 
 def test_find_modes_waveguide_dispersion() -> None:
-    modes = find_mode_dispersion(wg_width=0.45, resolution=20, cache=None)
+    modes = find_mode_dispersion(core_width=0.45, resolution=20, cache=None)
     m1 = modes
 
     # print(f"neff1 = {m1.neff}")